move loadscene to excalidraw-app and some refactoring
