+++ src/data/index.ts
@@ -24,28 +24,7 @@ import { canvasToBlob } from "./blob";
 export { loadFromBlob } from "./blob";
 export { saveAsJSON, loadFromJSON } from "./json";
 
-const BACKEND_GET = process.env.REACT_APP_BACKEND_V1_GET_URL;
-
 const BACKEND_V2_POST = process.env.REACT_APP_BACKEND_V2_POST_URL;
-const BACKEND_V2_GET = process.env.REACT_APP_BACKEND_V2_GET_URL;
-
-export const getImportedKey = (key: string, usage: KeyUsage) =>
-  window.crypto.subtle.importKey(
-    "jwk",
-    {
-      alg: "A128GCM",
-      ext: true,
-      k: key,
-      key_ops: ["encrypt", "decrypt"],
-      kty: "oct",
-    },
-    {
-      name: "AES-GCM",
-      length: 128,
-    },
-    false, // extractable
-    [usage],
-  );
 
 export const exportToBackend = async (
   elements: readonly ExcalidrawElement[],
@@ -100,52 +79,6 @@ export const exportToBackend = async (
   } catch (error) {
     console.error(error);
     window.alert(t("alerts.couldNotCreateShareableLink"));
-  }
-};
-
-const importFromBackend = async (
-  id: string | null,
-  privateKey?: string | null,
-): Promise<ImportedDataState> => {
-  try {
-    const response = await fetch(
-      privateKey ? `${BACKEND_V2_GET}${id}` : `${BACKEND_GET}${id}.json`,
-    );
-    if (!response.ok) {
-      window.alert(t("alerts.importBackendFailed"));
-      return {};
-    }
-    let data: ImportedDataState;
-    if (privateKey) {
-      const buffer = await response.arrayBuffer();
-      const key = await getImportedKey(privateKey, "decrypt");
-      const iv = new Uint8Array(12);
-      const decrypted = await window.crypto.subtle.decrypt(
-        {
-          name: "AES-GCM",
-          iv: iv,
-        },
-        key,
-        buffer,
-      );
-      // We need to convert the decrypted array buffer to a string
-      const string = new window.TextDecoder("utf-8").decode(
-        new Uint8Array(decrypted) as any,
-      );
-      data = JSON.parse(string);
-    } else {
-      // Legacy format
-      data = await response.json();
-    }
-
-    return {
-      elements: data.elements || null,
-      appState: data.appState || null,
-    };
-  } catch (error) {
-    window.alert(t("alerts.importBackendFailed"));
-    console.error(error);
-    return {};
   }
 };
 
@@ -251,25 +184,10 @@ export const exportCanvas = async (
   }
 };
 
-export const loadScene = async (
-  id: string | null,
-  privateKey: string | null,
-  // Supply initialData even if importing from backend to ensure we restore
-  // localStorage user settings which we do not persist on server.
-  // Non-optional so we don't forget to pass it even if `undefined`.
+export const loadScene = (
   initialData: ImportedDataState | undefined | null,
 ) => {
-  let data;
-  if (id != null) {
-    // the private key is used to decrypt the content from the server, take
-    // extra care not to leak it
-    data = restore(
-      await importFromBackend(id, privateKey),
-      initialData?.appState,
-    );
-  } else {
-    data = restore(initialData || {}, null);
-  }
+  const data = restore(initialData || {}, null);
 
   return {
     elements: data.elements,
